comment provider

+++ Patissier/Patissier/Product/ProductModel.h
@@ -12,9 +12,9 @@
     
 @property (strong, nonatomic) NSString *title;
 @property (strong, nonatomic) NSString *iD;
-@property (nonatomic ) NSInteger *price;
+@property (nonatomic) int *price;
 @property (strong, nonatomic) NSString *imageURLString;
     
-- (instancetype)initWithTitle:(NSString *)title iD:(NSString *)iD price:(NSInteger *)price;
+- (instancetype)initWithTitle:(NSString *)title iD:(NSString *)iD price:(int *)price;
     
 @end
